requirements.txt
# Core dependencies
langchain-anthropic>=0.1.0
langchain-core>=0.2.0
langchain>=0.2.0
anthropic>=0.18.0
pydantic>=2.0.0

# HTTP requests for API calls
requests>=2.31.0

# Database
asyncpg>=0.29.0

# FastAPI and server
fastapi>=0.104.1
uvicorn>=0.24.0

# Environment management
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
